backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv

langchain
langchain-community
langchain-openai      # or langchain-anthropic, etc.

langgraph
langchain-chroma
chromadb

pandas
numpy
tqdm

httpx